requirements.txt
boto3==1.35.5
botocore==1.35.5
jmespath==1.0.1
mysql-connector-python==9.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.2
